Move get_ptr/wrap_ptr to TypeSystem trait
